refactor: Token VO 패키지 이동
