+++ src/main/java/com/example/solidconnection/auth/domain/AccessToken.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.auth.service;
+package com.example.solidconnection.auth.domain;
 
 import com.example.solidconnection.siteuser.domain.Role;
