+++ src/main/java/com/example/solidconnection/auth/domain/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.auth.service;
+package com.example.solidconnection.auth.domain;
 
 public record RefreshToken(
         Subject subject,
