+++ src/main/java/com/example/solidconnection/auth/domain/Subject.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.auth.service;
+package com.example.solidconnection.auth.domain;
 
 public record Subject(
         String value
