+++ src/main/java/com/example/solidconnection/auth/dto/ReissueResponse.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.auth.dto;
 
-import com.example.solidconnection.auth.service.AccessToken;
+import com.example.solidconnection.auth.domain.AccessToken;
 
 public record ReissueResponse(
         String accessToken
